+++ test/testEmbedHandler.js
@@ -68,6 +68,12 @@ const logger = Logger.getTestLogger({
   level: 'info',
 });
 
+
+const crequest = {
+  extension: 'html',
+  url: '/test/test.html',
+};
+
 describe('Test Embed Handler', () => {
   it('Creates ESI', async () => {
     const node = {
@@ -93,6 +99,7 @@ describe('Integration Test with Embeds', () => {
     const result = await pipe(
       ({ content }) => ({ response: { status: 201, body: content.document.body.innerHTML } }),
       {
+        request: crequest,
         content: {
           body: `Hello World
 Here comes an embed.
@@ -115,6 +122,38 @@ https://www.youtube.com/watch?v=KOxbO0EI4MA
     assert.equal(result.response.body, `<p>Hello World
 Here comes an embed.</p>
 <esi:include src="https://example-embed-service.com/https://www.youtube.com/watch?v=KOxbO0EI4MA"></esi:include>
+<esi:remove><p><a href="https://www.youtube.com/watch?v=KOxbO0EI4MA">https://www.youtube.com/watch?v=KOxbO0EI4MA</a></p></esi:remove>
+<p><img src="easy.png" alt="Easy!" srcset="easy.png?width=480&amp;auto=webp 480w, easy.png?width=1384&amp;auto=webp 1384w, easy.png?width=2288&amp;auto=webp 2288w, easy.png?width=3192&amp;auto=webp 3192w, easy.png?width=4096&amp;auto=webp 4096w" sizes="100vw"></p>`);
+  });
+
+  it('html.pipe processes internal embeds', async () => {
+    const result = await pipe(
+      ({ content }) => ({ response: { status: 201, body: content.document.body.innerHTML } }),
+      {
+        request: crequest,
+        content: {
+          body: `Hello World
+Here comes an embed.
+
+./foo.md
+
+![Easy!](easy.png)
+`,
+        },
+      },
+      {
+        request: { params },
+        secrets,
+        logger,
+      },
+    );
+
+    assert.equal(result.response.status, 201);
+    assert.equal(result.response.headers['Content-Type'], 'text/html');
+    assert.equal(result.response.body, `<p>Hello World
+Here comes an embed.</p>
+<esi:include src="/test/foo.embed.html"></esi:include>
+<esi:remove><p>./foo.md</p></esi:remove>
 <p><img src="easy.png" alt="Easy!" srcset="easy.png?width=480&amp;auto=webp 480w, easy.png?width=1384&amp;auto=webp 1384w, easy.png?width=2288&amp;auto=webp 2288w, easy.png?width=3192&amp;auto=webp 3192w, easy.png?width=4096&amp;auto=webp 4096w" sizes="100vw"></p>`);
   });
 });