+++ test/testHTMLFromMarkdown.js
@@ -68,6 +68,11 @@ const logger = winston.createLogger({
   ],
 });
 
+const crequest = {
+  extension: 'html',
+  url: '/test/test.html',
+};
+
 /**
  * Assert that a specific html dom is generated from the given markdown
  * using our html pipeline.
@@ -86,7 +91,7 @@ const assertMd = async (md, html) => {
 
   const generated = await pipe(
     fromHTML,
-    { content: { body: multiline(md) } },
+    { content: { body: multiline(md) }, request: crequest },
     {
       logger,
       request: { params },
@@ -208,6 +213,30 @@ describe('Testing Markdown conversion', () => {
         Hello World [link](λ)
       `, `
         <h1 id="foo">Foo</h1>
+        <p>Hello World <a href="%CE%BB">link</a></p>
+    `);
+  });
+
+  it('HTML Block elements', async () => {
+    await assertMd(`
+        # Foo
+
+        <form>
+          <input type="text" name="name"><label for="name">Name</label>
+        </form>
+      `, `
+        <h1 id="foo">Foo</h1>
+        <form><input type="text" name="name"><label for="name">Name</label></form>
+    `);
+  });
+
+  it('HTML inline elements', async () => {
+    await assertMd(`
+        # Foo <em>Bar</em>
+
+        Hello World [link](λ)
+      `, `
+        <h1 id="foo-bar">Foo <em>Bar</em></h1>
         <p>Hello World <a href="%CE%BB">link</a></p>
     `);
   });